# remote write: use go 1.23 unique package to reduce duplication of interned labels when multiple remote writes are used