--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/prometheus/prometheus
 
-go 1.21.0
+go 1.23.0
 
 toolchain go1.22.5