--- a/storage/remote/intern_test.go
+++ b/storage/remote/intern_test.go
@@ -22,6 +22,7 @@ import (
 	"fmt"
 	"testing"
 	"time"
+	"unique"
 
 	"github.com/stretchr/testify/require"
 )
@@ -30,7 +31,7 @@ func TestIntern(t *testing.T) {
 	interner := newPool()
 	testString := "TestIntern"
 	interner.intern(testString)
-	interned, ok := interner.pool[testString]
+	interned, ok := interner.pool[unique.Make(testString)]
 
 	require.True(t, ok)
 	require.Equal(t, int64(1), interned.refs.Load(), fmt.Sprintf("expected refs to be 1 but it was %d", interned.refs.Load()))
@@ -41,13 +42,13 @@ func TestIntern_MultiRef(t *testing.T) {
 	testString := "TestIntern_MultiRef"
 
 	interner.intern(testString)
-	interned, ok := interner.pool[testString]
+	interned, ok := interner.pool[unique.Make(testString)]
 
 	require.True(t, ok)
 	require.Equal(t, int64(1), interned.refs.Load(), fmt.Sprintf("expected refs to be 1 but it was %d", interned.refs.Load()))
 
 	interner.intern(testString)
-	interned, ok = interner.pool[testString]
+	interned, ok = interner.pool[unique.Make(testString)]
 
 	require.True(t, ok)
 	require.Equal(t, int64(2), interned.refs.Load(), fmt.Sprintf("expected refs to be 2 but it was %d", interned.refs.Load()))
@@ -58,13 +59,13 @@ func TestIntern_DeleteRef(t *testing.T) {
 	testString := "TestIntern_DeleteRef"
 
 	interner.intern(testString)
-	interned, ok := interner.pool[testString]
+	interned, ok := interner.pool[unique.Make(testString)]
 
 	require.True(t, ok)
 	require.Equal(t, int64(1), interned.refs.Load(), fmt.Sprintf("expected refs to be 1 but it was %d", interned.refs.Load()))
 
 	interner.release(testString)
-	_, ok = interner.pool[testString]
+	_, ok = interner.pool[unique.Make(testString)]
 	require.False(t, ok)
 }
 
@@ -73,7 +74,7 @@ func TestIntern_MultiRef_Concurrent(t *testing.T) {
 	testString := "TestIntern_MultiRef_Concurrent"
 
 	interner.intern(testString)
-	interned, ok := interner.pool[testString]
+	interned, ok := interner.pool[unique.Make(testString)]
 	require.True(t, ok)
 	require.Equal(t, int64(1), interned.refs.Load(), fmt.Sprintf("expected refs to be 1 but it was %d", interned.refs.Load()))
 
@@ -84,7 +85,7 @@ func TestIntern_MultiRef_Concurrent(t *testing.T) {
 	time.Sleep(time.Millisecond)
 
 	interner.mtx.RLock()
-	interned, ok = interner.pool[testString]
+	interned, ok = interner.pool[unique.Make(testString)]
 	interner.mtx.RUnlock()
 	require.True(t, ok)
 	require.Equal(t, int64(1), interned.refs.Load(), fmt.Sprintf("expected refs to be 1 but it was %d", interned.refs.Load()))