@bboreham 2025-07-08
Hello from the bug-scrub!  @cstyan do you think you will come back to this?

Note that the interning hook was removed in #15456, as it was not compatible with the "stringlabels" implementation.